Add bluactive governor #1
Bluactive Governor https://forum.xda-developers.com/general/general/ref-to-date-guide-cpu-governors-o-t3048957 

Description:
A governor developed by eng.stk (presented in its Code_Blue cores) based on interactive with patches of ascending cafes and governor bits of on-demand type as well. This governor focuses primarily on performance as the other things that the developer creates, but it is also well balanced for the game and general use.