Enable USB Fast Charging #1
This is how SuperCharge works, the Huawei fast charge

SuperCharge, Huawei's fast charge, charges up to 58% of the
Huawei Hi6250 battery in just half an hour. 30 minutes
loading, a whole day enjoying!
Smart phone batteries, for better or for worse, have become
one of the main components when choosing a smartphone
as it depends largely on obtaining a satisfactory user experience.
Always be alert to the battery charge indicator, or limit the use made
of the smartphone is not something 
that favors the user experience, no matter how well the smartphone
behaves in the rest of its functions, It will be of little use
if Solo can use it quietly while having a nearby outlet.
To maximize battery life and enjoy your device for much longer without
worrying about the battery's charge status, 
leading smartphone manufacturers, such as Huawei, are developing
more energy-efficient components, obtaining in its terminals an exceptional performance, with a minimum impact on the life of the battery.
However, this is not always enough, so Huawei has also managed
to reduce the load times, so your smartphone will need to spend
less time connected to charge the battery. 30 minutes loading
a whole day enjoying!.